+++ crud/common/stash.lua
@@ -20,16 +20,12 @@ local stash = {}
 -- @tfield string stats_metrics_registry
 --  Stash for metrics rocks statistics registry.
 --
--- @tfield string select_module_compat_info
---  Stash for select compatability version registry.
---
 stash.name = {
     cfg = '__crud_cfg',
     stats_internal = '__crud_stats_internal',
     stats_local_registry = '__crud_stats_local_registry',
     stats_metrics_registry = '__crud_stats_metrics_registry',
     ddl_triggers = '__crud_ddl_spaces_triggers',
-    select_module_compat_info = '__select_module_compat_info',
     storage_readview = '__crud_storage_readview',
 }