+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Drop support of Tarantool 1.10.x older than 1.10.8 and
   Tarantool 2.x older than 2.5.2 (#365).
 * Unlock `tuple-keydef` and `tuple-merger` version dependencies (#365).
+* Require `tuple-keydef` and `tuple-merger` for work (#365).
 
 ## [1.4.0] - 16-10-23